+++ test/channels.jl
@@ -44,12 +44,6 @@ end
 end
 
 @testset "Task constructors" begin
-    c = Channel(ctype=Float32, csize=2) do c; map(i->put!(c,i), 1:100); end
-    @test eltype(c) == Float32
-    @test c.sz_max == 2
-    @test isopen(c)
-    @test collect(c) == 1:100
-
     c = Channel{Int}() do c; map(i->put!(c,i), 1:100); end
     @test eltype(c) == Int
     @test c.sz_max == 0
@@ -59,18 +53,28 @@ end
     @test c.sz_max == 0
     @test collect(c) == [1, "hi"]
 
+    c = Channel(Inf) do c; put!(c,1); end
+    @test eltype(c) == Any
+    @test c.sz_max == typemax(Int)
     c = Channel{Int}(Inf) do c; put!(c,1); end
     @test eltype(c) == Int
     @test c.sz_max == typemax(Int)
 
     taskref = Ref{Task}()
-    c = Channel{Int}(csize=0, taskref=taskref) do c; put!(c, 0); end
+    c = Channel{Int}(0, taskref=taskref) do c; put!(c, 0); end
     @test eltype(c) == Int
     @test c.sz_max == 0
     @test istaskstarted(taskref[])
     @test !istaskdone(taskref[])
     take!(c); wait(taskref[])
     @test istaskdone(taskref[])
+
+    # Legacy constructor
+    c = Channel(ctype=Float32, csize=2) do c; map(i->put!(c,i), 1:100); end
+    @test eltype(c) == Float32
+    @test c.sz_max == 2
+    @test isopen(c)
+    @test collect(c) == 1:100
 end
 
 @testset "multiple concurrent put!/take! on a channel for different sizes" begin
@@ -230,7 +234,7 @@ using Distributed
 
     for T in [Any, Int]
         taskref = Ref{Task}()
-        chnl = Channel(tf6, ctype=T, csize=N, taskref=taskref)
+        chnl = Channel{T}(tf6, N, taskref=taskref)
         put!(chnl, 2)
         yield()
         @test_throws ErrorException wait(chnl)